add long long int

+++ lab31proxy/Constants.h
@@ -12,7 +12,7 @@ namespace ProxyServer {
     static const int BUF_SIZE = 1024;
     static int DEFAULT_PORT = 80;
     static int COUNT_CASH_ELEMENT = 100;
-    static int SIZE_EACH_CASH_ELEMENT = 1024 * 1024;
+    static long long int SIZE_EACH_CASH_ELEMENT = 1024 * 1024;
     static int MAX_CASH_SIZE = 1024 * 1024; // 1 кб
 
     static std::string REGEX_FOR_HOST_AND_PORT = "Host: ([\\w.\\d-]*)(:(\\d)+)*";
